changed fallback macro to have the same return type as rint().
2008-05-22  Jan Heller  <jheller@svn.gnome.org>

        * extensions/util.h: changed fallback macro to have the
        same return type as rint().

svn path=/trunk/; revision=315

+++ extensions/util.h
@@ -25,7 +25,7 @@
 #include "config.h"
 
 #ifndef HAVE_RINT
-# define rint(f)  ((long) floor (((double) f) + 0.5))
+# define rint(f)  (floor (((double) (f)) + 0.5))
 #endif
